requirements.txt
gradio
torch
torchvision
numpy
Pillow
opencv-python
tqdm